Update frontend models
