--- frontend/src/models/room.ts
@@ -1,8 +1,0 @@
-import Video from "./video";
-
-export default interface Room {
-  name: string;
-  currVideoTitle: string;
-  currVideoId: string;
-  videoQueue: Video[];
-}
