--- frontend/src/models/video.ts
@@ -1,7 +1,0 @@
-export default interface Video {
-  id: string;
-  title: string;
-  youtubeId: string;
-  channel: string;
-  lengthInSeconds: number;
-}
