--- frontend/src/models/videoState.ts
@@ -1,13 +1,0 @@
-export enum PlayerState {
-  UNSTARTED = -1,
-  ENDED = 0,
-  PLAYING = 1,
-  PAUSED = 2,
-  BUFFERING = 3,
-  CUED = 5
-}
-
-export interface VideoState {
-  secondsElapsed: number;
-  playerState: PlayerState;
-}
